requirements.txt
Flask==3.0.3 
python-dotenv==1.0.1 
requests==2.32.3 
amadeus==9.0.0 
pytz==2024.1 
python-dateutil==2.9.0.post0 
openai==1.51.2 
httpx==0.27.2
